Q5 Q6 sans correction

+++ src/controllers/author.controller.ts
@@ -3,6 +3,7 @@ import { authorService } from "../services/author.service";
 import { AuthorDTO } from "../dto/author.dto";
 import { Author } from "../models/author.model";
 import { CustomError } from "../middlewares/errorHandler";
+import { Book } from "../models/book.model";
 
 @Route("authors")
 @Tags("Authors")
@@ -17,7 +18,7 @@ export class AuthorController extends Controller {
   @Get("{id}")
   public async getAuthorById(@Path() id: number): Promise<AuthorDTO> {
     let author: Author | null = await authorService.getAuthorById(id);
-    if(author === null) {
+    if (author === null) {
       let error: CustomError = new Error("Author not found");
       error.status = 404;
       throw error;
@@ -32,7 +33,7 @@ export class AuthorController extends Controller {
     @Body() requestBody: AuthorDTO
   ): Promise<AuthorDTO> {
     const { firstName, lastName } = requestBody;
-    if(!firstName || !lastName) {
+    if (!firstName || !lastName) {
       let error: CustomError = new Error("First name and last name are required");
       error.status = 400;
       throw error;
@@ -43,6 +44,12 @@ export class AuthorController extends Controller {
   // Supprime un auteur par ID
   @Delete("{id}")
   public async deleteAuthor(@Path() id: number): Promise<void> {
+    let hasCopy = await authorService.hasBookCopy(id)
+    if (hasCopy) {
+      let error: CustomError = new Error("This author have a book copy");
+      error.status = 400;
+      throw error;
+    }
     await authorService.deleteAuthor(id);
   }
 
@@ -54,14 +61,14 @@ export class AuthorController extends Controller {
   ): Promise<AuthorDTO> {
     const { firstName, lastName } = requestBody;
 
-    if(!firstName || !lastName) {
+    if (!firstName || !lastName) {
       let error: CustomError = new Error("First name and last name are required");
       error.status = 400;
       throw error;
     }
 
     let author = await authorService.updateAuthor(id, firstName, lastName)
-    if(author === null){
+    if (author === null) {
       let error: CustomError = new Error("Author not found");
       error.status = 404;
       throw error;
@@ -69,4 +76,25 @@ export class AuthorController extends Controller {
       return author;
     }
   }
+
+  // Récupère la liste des livres de l'auteur
+  // @Get("{id}/books")
+  // public async getBooksAuthor(@Path() id: number): Promise<Book[]> {
+  //   // Vérifie que l'auteur existe
+  //   let author: Author | null = await authorService.getAuthorById(id);
+  //   if (author === null) {
+  //     let error: CustomError = new Error("Author not found");
+  //     error.status = 404;
+  //     throw error;
+  //   }
+
+  //   let books: Book[] | null = await authorService.getBooksAuthor(id);
+  //   if (books === null) {
+  //     let error: CustomError = new Error("Author have no books");
+  //     error.status = 404;
+  //     throw error;
+  //   } else {
+  //     return books;
+  //   }
+  // }
 }
